+++ docs/changelog.md
@@ -3,6 +3,7 @@
 ## Unreleased
 * Minimum required Gradle version is now 9.1
 * Fixed support for Android Gradle Plugin version 9.0.1
+* Added `com.osacky.fulladle.settings` settings plugin for Gradle configuration cache compatible multi-module testing
 
 ## 0.19.0
 * Minimum required JVM version is now 17.